Updated DatabaseManager
Created some function for DatabaseManager. Also need to test them.
Also created a 'Testing' folder in which I'm planning to put all of the test
files, so I'll need to refactor the paths.

Next is DatabaseAdapter, which is will communicate with MusicManager in order
to add or fetch songs/genres etcetera.