Created HouseMusic (main class) + small updates
Created HouseMusic class, which is the main class containing all dictionaries and configurations.
Imported python-Levenshtein in order to use it inside VoiceCommandAdapter s.t. I can measure
the distance between strings and pick the function that is closest to the recognized command.

I'm planning to create a VoiceListener class that will start a separate thread dedicated
to listening. That specific thread will live inside HouseMusic and will make use
of VoiceCommandAdapter (who also uses VoiceRecognition class) in order to
execute the commands.