Updated schema & added repo DP
Updated the database schema & the corresponding sqlachemy ORM models (Song - repository design pattern).

We are now storing multiple values related to a song, such as play time, average play time, play count, all expressed in minutes.

Defined the concept of a "play/listening cycle" as being the arbitrarly chosen time interval of user activity. Currently it lasts a week. Essentialy, it's a variable used to define the "present". I am using this in order to compare the present listening experience to the past experience and, based on that, to decide whether to keep a song or not.