Added chronjob to check bluetooth connection
Decided to use the Flask APScheduler instead of threads. Created a job that executes every 15 seconds and checks the bluetooth connection. If the bluetooth device is not connected, it attempts to connect. If connection is failing, it will try again after 15 seconds.

Regardless of the result, the job emits a socket message to all the clients informing them about the bluetooth connection status.